# swagger.yaml
openapi: 3.0.0
info:
  title: Saint-Pierre API
  version: 1.0.0
  description: API complète pour la gestion des devoirs, cours et messagerie scolaire
  contact:
    name: Support
  x-logo:
    url: https://your-api-logo.png
servers:
  - url: http://localhost:3000
    description: Serveur de développement
  - url: https://api.saint-pierre.com
    description: Serveur de production
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /auth/sign-in
  schemas:
...
    Utilisateur:
      type: object
      properties:
        id_user:
          type: string
          pattern: '^[0-9]+$'
          example: '1'
        nom:
          type: string
          maxLength: 50
        prenom:
          type: string
          maxLength: 50
        login:
          type: string
          maxLength: 50
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/Role'
        created_at:
          type: string
          format: date-time
        eleve:
          $ref: '#/components/schemas/Eleve'
        professeur:
          $ref: '#/components/schemas/Professeur'
    Role:
      type: string
      enum: [eleve, professeur, administrateur]
    Eleve:
      type: object
      properties:
        id_user:
          type: string
          pattern: '^[0-9]+$'
        id_classe:
          type: string
          pattern: '^[0-9]+$'
          nullable: true
        annee:
          type: string
          nullable: true
        classe:
          $ref: '#/components/schemas/Classe'
        specialites:
          type: array
          items:
            $ref: '#/components/schemas/Specialite'
        options:
          type: array
          items:
            $ref: '#/components/schemas/Option'
    Professeur:
      type: object
      properties:
        id_user:
          type: string
          pattern: '^[0-9]+$'
        matiere:
          type: string
          maxLength: 50
          nullable: true
        classes_enseignees:
          type: array
          items:
            $ref: '#/components/schemas/Classe'
        specialites_enseignees:
          type: array
          items:
            $ref: '#/components/schemas/Specialite'
        options_enseignees:
          type: array
          items:
            $ref: '#/components/schemas/Option'

    # ========== CLASSE & STRUCTURES SCOLAIRES ==========
    Classe:
      type: object
      properties:
        id_classe:
          type: string
          pattern: '^[0-9]+$'
          example: '1'
        nom_classe:
          type: string
          maxLength: 50
          example: 'Terminale A'
        niveau:
          type: string
          maxLength: 20
          example: 'Terminale'
        lettre:
          type: string
          maxLength: 1
          example: 'A'
        created_at:
          type: string
          format: date-time
...
    Specialite:
      type: object
      properties:
        id_specialite:
          type: string
          pattern: '^[0-9]+$'
        nom_specialite:
          type: string
          maxLength: 50
          example: 'NSI'
        created_at:
          type: string
          format: date-time
    Option:
      type: object
      properties:
        id_option:
          type: string
          pattern: '^[0-9]+$'
        nom_option:
          type: string
          maxLength: 50
          example: 'Latin'
        created_at:
          type: string
          format: date-time

    # ========== COURS ==========
...
    Error400:
      type: object
      properties:
        error:
          type: string
          example: 'Paramètres invalides'
    Error401:
      type: object
      properties:
        error:
          type: string
          example: 'Accès refusé: token manquant'
    Error403:
      type: object
      properties:
        error:
          type: string
          example: 'Accès refusé: rôle insuffisant'
    Error404:
      type: object
      properties:
        error:
          type: string
          example: 'Ressource non trouvée'
    Error500:
      type: object
      properties:
        error:
          type: string
          example: 'Erreur serveur interne'
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
          example: 'Opération réussie'
...
security:
  - BearerAuth: []
paths:
...
  /auth/sign-up:
    post:
      summary: Créer un nouveau compte
      tags: [Auth]
      security: []
      operationId: signUp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nom:
                  type: string
                  minLength: 1
                  maxLength: 50
                prenom:
                  type: string
                  minLength: 1
                  maxLength: 50
                login:
                  type: string
                  minLength: 1
                  maxLength: 50
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 6
                classe:
                  type: string
                  pattern: '^[0-9]+$'
                annee:
                  type: string
                role:
                  $ref: '#/components/schemas/Role'
              required: [nom, prenom, login, email, password, classe, annee]
      responses:
        '201':
          description: Compte créé avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  id_user:
                    type: string
                  message:
                    type: string
        '400':
          description: Données manquantes ou invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'
  /auth/sign-in:
    post:
      summary: Se connecter
      tags: [Auth]
      security: []
      operationId: signIn
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                login:
                  type: string
                password:
                  type: string
              required: [login, password]
      responses:
        '200':
          description: Connexion réussie
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: 'eyJhbGciOiJIUzI1NiIs...'
                  id_user:
                    type: string
                  role:
                    $ref: '#/components/schemas/Role'
        '401':
          description: Identifiants invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error401'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'
  /auth/me:
    get:
      summary: Récupérer le profil actuel
      tags: [Auth]
      operationId: getCurrentUser
      responses:
        '200':
          description: Profil utilisateur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Utilisateur'
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error401'
        '403':
          description: Token expiré
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error403'
  /auth/forgot-password:
    post:
      summary: Demander une réinitialisation de mot de passe
      tags: [Auth]
      security: []
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
              required: [email]
      responses:
        '200':
          description: Email de réinitialisation envoyé
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: Utilisateur non trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
  /auth/reset-password:
    post:
      summary: Réinitialiser le mot de passe
      tags: [Auth]
      security: []
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reset_token:
                  type: string
                new_password:
                  type: string
                  minLength: 6
              required: [reset_token, new_password]
      responses:
        '200':
          description: Mot de passe réinitialisé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Token invalide ou expiré
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400'
...